+++ CHANGES.rst
@@ -12,7 +12,7 @@ Changes
 Features / Changes
 ~~~~~~~~~~~~~~~~~~~~~
 * Add ``PATCH /permissions`` endpoint that updates permissions and creates related resources if necessary.
-* ``permissions.cfg`` now supports a new format for the ``type`` parameter, using multiple types separated
+* Add support of new format for ``permissions.cfg`` for the ``type`` parameter, using multiple types separated
   by a slash character, matching each type with each resource found in the ``resource`` parameter.
 
 .. _changes_3.23.0: